Setting scroll motion value to default 0 when page loads
When page loads, the containerRef is not assigned, leading the scrollYProgress to have no reference and scrollYProgress getting set to 1. This is smoothed out by the smoothYProgress spring, which makes the page scroll down to 1 on load. To fix, setting the scrollYProgress to motionValue of 0 owhen the containerRef is undefined